Remove explicit `setup` task and fix up tasks that were depending on it incorrectly
