+++ src/main/groovy/org/labkey/gradle/task/DeployAppBase.groovy
@@ -1,7 +1,6 @@
 package org.labkey.gradle.task
 
 import org.apache.commons.lang3.SystemUtils
-import org.gradle.api.DefaultTask
 import org.gradle.api.file.ConfigurableFileCollection
 import org.gradle.api.file.CopySpec
 import org.gradle.api.file.DirectoryProperty
@@ -13,7 +12,7 @@ import org.labkey.gradle.util.BuildUtils
 
 import javax.inject.Inject
 
-abstract class DeployAppBase extends DefaultTask {
+abstract class DeployAppBase extends SetUpProperties {
 
     @Inject abstract FileSystemOperations getFs()
